+++ CHANGELOG.md
@@ -1,6 +1,28 @@
 Changelog
 ===
 
+## Version 15.7.3
+### 🛠 Fixes & Updates
+
+* **callout:** space margin-less blocks like tables inside callouts ([#1627](https://github.com/readmeio/markdown/issues/1627)) ([5668182](https://github.com/readmeio/markdown/commit/5668182667ba3ec7d47a835552e15139d7b184c4))
+
+## Version 15.7.2
+### 🛠 Fixes & Updates
+
+* **callout:** keep list-leading body beside the floated icon ([#1624](https://github.com/readmeio/markdown/issues/1624)) ([fd64bad](https://github.com/readmeio/markdown/commit/fd64bad90c45649eb43b61f78453bf7bfd161f99)), closes [#825](https://github.com/readmeio/markdown/issues/825)
+
+## Version 15.7.1
+### 🛠 Fixes & Updates
+
+* **mdxish:** keep components under list items claimed across lazy lines ([#1621](https://github.com/readmeio/markdown/issues/1621)) ([1a7e823](https://github.com/readmeio/markdown/commit/1a7e823d4af206ba324424153f2d32e577f2581f))
+* **MdxishEditor:** self closing components under a list having its closing tag trimmed  ([#1622](https://github.com/readmeio/markdown/issues/1622)) ([0e1e8fd](https://github.com/readmeio/markdown/commit/0e1e8fdb78bbb0392ce3327d39c6a91663e050c4)), closes [#1621](https://github.com/readmeio/markdown/issues/1621)
+
+## Version 15.7.0
+### ✨ New & Improved
+
+* **MdxishEditor:** retain lowercase <table> syntax during serialization ([#1620](https://github.com/readmeio/markdown/issues/1620)) ([7457885](https://github.com/readmeio/markdown/commit/7457885ebe103fac32919d1cd453920becf9204c))
+* show three heading levels in the table of contents ([#1615](https://github.com/readmeio/markdown/issues/1615)) ([474c005](https://github.com/readmeio/markdown/commit/474c005934c952ce46afea6f4a4cb8c62f200d24))
+
 ## Version 15.6.0
 ### ✨ New & Improved